Refactor db connection in main
